backend/requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# Pydantic for data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# YAML configuration
PyYAML>=6.0

# HTTP requests and API calls
httpx>=0.24.0
requests>=2.31.0

# Database and storage
supabase>=2.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Notion API
notion-client>=2.0.0

# Web scraping and content extraction
beautifulsoup4>=4.12.0
scrapy>=2.11.0
selenium>=4.15.0

# AI and ML libraries
openai>=1.0.0
anthropic>=0.7.0
transformers>=4.35.0
torch>=2.1.0

# Image generation and processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Audio processing
pydub>=0.25.1
librosa>=0.10.0

# Video processing
moviepy>=1.3.3

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
